Use double instead of float

+++ gidpixhi/math.hh
@@ -1,9 +1,9 @@
 #pragma once
 #include <cmath>
 
-// TODO: allow any numeric type in place of 'float'
+// TODO: allow any numeric type in place of 'double'
 struct Point {
-	float x {}, y {}, z {};
+	double x {}, y {}, z {};
 	auto operator<=>(Point const&) const = default;
 };
 
@@ -13,21 +13,21 @@ Point sub(Point a, Point b) {
 	return {a.x-b.x, a.y-b.y, a.z-b.z};
 }
 
-float dot(Point a, Point b) {
+double dot(Point a, Point b) {
 	return a.x*b.x + a.y*b.y + a.z*b.z;
 }
 
-float distanceSquared(Point a, Point b) {
+double distanceSquared(Point a, Point b) {
 	Point d = sub(a, b);
 	return d.x*d.x + d.y*d.y + d.z*d.z;
 }
 
-float distance(Point a, Point b) {
+double distance(Point a, Point b) {
 	return std::sqrt(distanceSquared(a, b));
 }
 
 Point norm(Point v) {
-	float dist = distance(v, {0, 0, 0});
+	double dist = distance(v, {0, 0, 0});
 	return {v.x/dist, v.y/dist, v.z/dist};
 }
